\name{cov2corr}
\alias{cov2corr}
\title{make a correlation matrix from a covariance matrix}
\usage{
cov2corr(m)
}
\arguments{
  \item{m}{a covariance matrix}
}
\description{
If m is a covariance matrix, this returns the corresponding
correlation matrix.
}
\note{
There is a similar function in base R or a similar name,
but Hans still calls this in places.
}
\author{
Hans J. Skaug \email{Hans.Skaug@math.uib.no}
}

