\name{member}
\alias{member}
\title{test each element of x to see if it is in y}
\usage{
member(x, y)
}
\arguments{
  \item{x}{a vector}

  \item{y}{another vector}
}
\description{
A function that delivers the same functionality as x %in% y
}
\note{
There is similar functionality in base R, but Hans calls
this in places so we include it here.
}
\author{
Hans J. Skaug \email{Hans.Skaug@math.uib.no}
}

