\docType{package}
\name{psiblings-package}
\alias{psiblings}
\alias{psiblings-package}
\title{Infers siblings groups from genetic data by pairwise and graphical computations}
\description{
Infers siblings groups from genetic data by pairwise and
graphical computations.
}
\details{
\tabular{ll}{ Package: \tab psiblings\cr Type: \tab
Package\cr Version: \tab 0.1\cr Date: \tab 2008-08-24\cr
License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

Using the Fermat primality test, pseudoprime checks for
primes probabilistically; the test is fooled every time by
Carmichael numbers. \code{\link{is.pseudoprime}} checks a
number \code{n} for pseudoprimality, applying Fermat's test
\code{times} times.
}
\examples{
x <- 1:10  # have to put more in here.
}
\author{
Hans J. Skaug and Eric C. Anderson
\email{Hans.Skaug@mi.uib.no} and
\email{eric.anderson@noaa.gov}
}
\references{
None Yet
}
\keyword{package}

