\name{readH}
\alias{readH}
\title{read Hessian of dimension n from admb .cor file}
\usage{
readH(file, n, cor = F)
}
\arguments{
  \item{file}{path to file holding the Hessian}

  \item{n}{number of rows and columns of the Hessian}

  \item{cor}{logical indicating whether to return...}
}
\description{
read Hessian of dimension n from admb .cor file
}
\author{
Hans J. Skaug \email{Hans.Skaug@math.uib.no}
}

