\name{std_read}
\alias{std_read}
\title{Read from a .std file produced by admb}
\usage{
std_read(name)
}
\arguments{
  \item{name}{path of file to read from}
}
\description{
Read from a .std file produced by admb or a file of
identical structure.
}
\author{
Hans J. Skaug \email{Hans.Skaug@math.uib.no}
}

